docs/requirements.txt
sphinx>=4.2.0
sphinx_rtd_theme>=1.0.0
ipython>=7.4
ipykernel>=5.1.0
numpydoc>=1.1.0
recommonmark==0.5.0
sphinx-gallery
GitPython
mvlearn
